+++ CHANGELOG.md
@@ -1,5 +1,9 @@
-# 0.10 (unreleased)
+# 0.11 (unreleased)
+
 **Goal: Offer actions when runtime errors occur.**
+**Goal: Evaluate functions up to the cursor.**
+
+# 0.10 (released 30th November 2023)
 
 **Goal: Warn on loading functions with free variables.**